pyproject.toml
[project]
name = "{{project_name}}"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-model-utils>=5.0.0",
    "django-permissions-policy>=4.26.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.9",
    "python-decouple>=3.8",
    "whitenoise[brotli]>=6.9.0",
]


[dependency-groups]
dev = [
    "model-bakery>=1.20.4",
]
